Move plugins/manual.py to _internal/
